+++ intellij-plugin/src/main/kotlin/com/apollographql/ijplugin/settings/SettingsComponent.kt
@@ -1,11 +1,8 @@
 package com.apollographql.ijplugin.settings
 
 import com.apollographql.ijplugin.ApolloBundle
-import com.apollographql.ijplugin.lsp.isLspAvailable
 import com.apollographql.ijplugin.project.apolloProjectService
-import com.apollographql.ijplugin.settings.lsp.lspGroup
 import com.apollographql.ijplugin.settings.studio.ApiKeyDialog
-import com.intellij.openapi.observable.properties.GraphProperty
 import com.intellij.openapi.observable.properties.PropertyGraph
 import com.intellij.openapi.project.Project
 import com.intellij.ui.AddEditRemovePanel
@@ -16,13 +13,11 @@ import javax.swing.JCheckBox
 import javax.swing.JComponent
 import javax.swing.JPanel
 
-
 class SettingsComponent(private val project: Project) {
   private val propertyGraph = PropertyGraph()
   private val automaticCodegenTriggeringProperty = propertyGraph.property(false)
   private val contributeConfigurationToGraphqlPluginProperty = propertyGraph.property(false)
   private val telemetryEnabledProperty = propertyGraph.property(false)
-  private val lspModeEnabledProperty: GraphProperty<Boolean> = propertyGraph.property(false)
 
   var automaticCodegenTriggering: Boolean by automaticCodegenTriggeringProperty
   var contributeConfigurationToGraphqlPlugin: Boolean by contributeConfigurationToGraphqlPluginProperty
@@ -32,7 +27,6 @@ class SettingsComponent(private val project: Project) {
       addEditRemovePanel?.data = value.toMutableList()
     }
   var telemetryEnabled: Boolean by telemetryEnabledProperty
-  var lspModeEnabled: Boolean by lspModeEnabledProperty
 
   private lateinit var chkAutomaticCodegenTriggering: JCheckBox
   private var addEditRemovePanel: AddEditRemovePanel<ApolloKotlinServiceConfiguration>? = null
@@ -104,10 +98,6 @@ class SettingsComponent(private val project: Project) {
               .comment(ApolloBundle.message("settings.studio.apiKeys.comment"))
         }
       }
-    }
-
-    if (isLspAvailable()) {
-      lspGroup(lspModeEnabledProperty)
     }
 
     group(ApolloBundle.message("settings.telemetry.telemetryEnabled.title")) {
