+++ intellij-plugin/src/main/resources/messages/ApolloBundle.properties
@@ -121,7 +121,7 @@ settings.rover.checking=Checking for Rover...
 settings.rover.installed.withLsp=<html>Rover <b>{0}</b> is installed. ✅
 settings.rover.installed.noLsp=<html><font color="#FF8080">Rover <b>{0}</b> is installed, please install version <b>0.27.0</b> or more recent.
 settings.rover.notInstalled=<html><font color="#FF8080">Rover doesn\'t seem to be installed.
-settings.rover.notInstalled.instructions=Please refer to the <a href=\"https://www.apollographql.com/docs/rover/getting-started\">installation instructions</a>.<br>⚠️Please make sure to install version <b>0.27.0</b> or more recent.
+settings.rover.notInstalled.instructions=Please refer to the <a href=\"https://www.apollographql.com/docs/rover/getting-started\">installation instructions</a>. Please make sure to install version <b>0.27.0</b> or more recent.
 settings.rover.lsp.disabled=Disabled
 settings.rover.lsp.disabled.help=Disable the Rover integration. The schema editing experience will rely on the GraphQL plugin's implementation. Recommended for Apollo Kotlin users.
 settings.rover.lsp.enabled=Enabled
